KAFKA-14274: Introduce FetchRequestManager to integrate fetch into new consumer threading refactor
This change introduces the FetchRequestManager that will be
responsible for:

- Formatting fetch requests to the background thread
- Configuring the callback on fetch responses for the background
  thread

The response handler will collect the fetch responses from the broker
and create CompletedFetch instances as is done in Fetcher. The newly
The foreground logic will decompress the data into a Record, which
will then be deserialized into a ConsumerRecord for returning to the user.